Add stricter checks for the length of additional filter lists
If the list of additional filters is nonempty, enforce its length to match
the number of arguments of the CAP operation. This will prevent that one
forgets about the filter lists when updating a category to accept the
category as the first argument of the added functions in the next commit.