TODO: Assert that Range and Source are equal in PreCompose
See also the comment issue 16 in LazyCategories:

For future reference: I have tried to do type checking during compilation (i.e. compare the syntax trees of the range of the pre morphism and the source of the post morphism in PreCompose) and in principle this works. However, there are lots of false positives because

    the compiler does not know about equalities in the specifications (e.g. Range( alpha ) = Source( beta ) in the specification of HomologyObject) and
    syntax trees can only be properly compared after everything is resolved and inlined. Currently, the compiler makes no guarantees about when this property is true for some subtree. Thus, the compiler sometimes compares partially resolved/inlined subtrees which usually are not equal as syntax trees.

For solving the first point, this equalities have to be given in a way readable for the compiler. For solving the second point I can imagine collecting all required equalities during compilation and defer the type checking until the end of the compilation when everything is guaranteed to be resolved and inlined.